Clean up Vector
* Match types in vector_insert and vector_delete (int -> size_t)
* Removed unneeded (void *)
* Removed unneeded assignment

+++ vector/vector.h
@@ -15,8 +15,8 @@ Vector *vector_create();
 void vector_free(Vector *vector);
 
 int vector_add(Vector *vector, void *item);
-int vector_insert(Vector *vector, void *item, int index);
-int vector_delete(Vector *vector, int index);
+int vector_insert(Vector *vector, void *item, size_t index);
+int vector_delete(Vector *vector, size_t index);
 
 
 #endif // CDSLIB_VECTOR_H
